Automatically update the README with a table of all RFCs/ADRs
This commit adds a script that updates the repo root's README.md file
with a table of contents of all ADRs and RFCs it finds in the repo.

A GitHub Actions workflow accompanying the script is added that
runs the script and creates a PR to update the README. The workflow is
run on each merge to main but will only create a PR if any changes to
the README are actually applied.