Add a section and TODO re Manual Approval
There will be some little bits of mechanism that will need to be
adapted for the proposed architecture.

This commit makes space to explain how the approval mechanism should
work in the proposed architecture. I've filled in an explanation of
how it works now, and left a TODO note for figuring out how it _could_
work, to do once we get the proposed architecture straight.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>