Explain the promotion algorithm
Here I've elaborated on the "promotion algorithm", or how the
controller decides what to do next. I left some TODOs where I think
some detail needs further explanation.

This includes a subtle change: I have left out the possibility of
triggering promotions when _just one_ target application has changed
state. This is the way it works today because that's the only way it
can work with notifications and no state; but it's not a mode of
operation I would expect as the default, or want to use.

Signed-off-by: Michael Bridgen <michael.bridgen@weave.works>